[move] Remove Move IR `Transaction.Foo` syntax
Move IR allows for referring to the current "transaction" address when
identifying a module, using the special identifier `Transaction`, as in
`Transaction.<module-name>`. The current transaction's address was
typically passed in by the user compiling the Move IR, on the
command line.

Now that the Move IR `compiler` executable no longer supports a
command-line argument to specify an `--address`, the `Transaction`
syntax is more obscure: it could only be used when declaring imports or
friends of a module with an address. Attempting to
`import Transaction.Foo` within a Move IR script, which can no longer
be given an address, would result in an error.

Because it is (and can be) rarely used, remove the `Transaction.Foo`
syntax altogether.