[move] Remove Move IR type inference
Simplify the Move IR language and compiler by eliminating the need to
keep track of the types of expressions.

For the most part, this requires no change to the language, except for one
aspect: when struct fields are borrowed, the user must now explicitly
specify which struct type is being accessed. This is because, for
example, a simple field identifier `f` may refer to different fields in
any number of different structs `S1`, `S2`, etc.

Closes: #9040