[move-prover] an analysis pass for the global invariants
This pass collects information on how to inject global invariants into
the bytecode of a function.

The end result of this analysis is summarized in two structs:

```rust
/// A named struct for holding the information on how an invariant is
/// relevant to a bytecode location.
struct PerBytecodeRelevance {
    /// for each `inst_fun` (instantiation of function type parameters) in the key set, the
    /// associated value is a set of `inst_inv` (instantiation of invariant type parameters) that
    /// are applicable to the concrete function instance `F<inst_fun>`.
    insts: BTreeMap<Vec<Type>, BTreeSet<Vec<Type>>>,
}

/// A named struct for holding the information on how invariants are relevant to a function.
struct PerFunctionRelevance {
    /// Invariants that needs to be assumed at function entrypoint
    /// - Key: global invariants that needs to be assumed before the first instruction,
    /// - Value: the instantiation information per each related invariant.
    entrypoint_assumptions: BTreeMap<GlobalId, PerBytecodeRelevance>,

    /// For each bytecode at given code offset, the associated value is a map of
    /// - Key: global invariants that needs to be asserted after the bytecode instruction and
    /// - Value: the instantiation information per each related invariant.
    per_bytecode_assertions: BTreeMap<CodeOffset, BTreeMap<GlobalId, PerBytecodeRelevance>>,
}
```

A note about `PerBytecodeRelevance`: in fact, in this phase, we don't intend
to instantiation the function nor do we want to collect information on how this
function (or this bytecode) needs to be instantiated. All we care is how the
invariant should be instantiated in order to be instrumented at this code point,
with a generic unction and generic code.

But unfortunately, based on how the type unification logic is written now,
this two-step instantiation is needed in order to find all possible
instantiations of the invariant. I won't deny that there might be a way to
collect invariant instantiation combinations without instantiating the
function type parameters, but I haven't iron out one so far.

Closes: #9072